+++ src/data/effect-handlers/alteration.ts
@@ -1,5 +1,6 @@
 import { BaseMaterialEffect } from "../material";
 import { RefinedItem } from "@src/refined-item";
+import { OwnedMaterial } from "@src/material";
 
 export type AlterationEffect = BaseMaterialEffect & {
     type: "Alteration";
@@ -10,18 +11,20 @@ export type AlterationEffect = BaseMaterialEffect & {
 
 export class AlterationEffectHandler {
     static async handleUpdate({
-        item,
+        material,
         effect,
         changes,
     }: {
-        item: RefinedItem;
+        material: OwnedMaterial;
         effect: AlterationEffect;
         changes: Record<string, unknown>;
     }) {
         if (effect.editData) return;
         if (
-            typeof foundry.utils.getProperty(item.item, effect.property) !==
-            "undefined"
+            typeof foundry.utils.getProperty(
+                material.parent.item,
+                effect.property,
+            ) !== "undefined"
         ) {
             foundry.utils.mergeObject(
                 changes,